GeometricSolutions

Documentation for GeometricSolutions.

GeometricSolutions.EnsembleSolution

GeometricSolutions.GeometricSolution

GeometricSolutions.EnsembleSolution — Type

EnsembleSolution: Collection of all solutions of an EnsembleProblem.

GeometricSolutions.GeometricSolution — Type

GeometricSolution: Solution of a geometric differential equation

Contains all fields necessary to store the solution of a GeometricProblem.

Fields

t: time steps

s: NamedTuple of DataSeries for each solution component

step: store every step'th time step (default: 1)

nstore: number of time steps to store

offset: offset of current time step

Constructors

GeometricSolution(problem, step = 1)

The usual way to initialise a Solution is by passing a GeometricEquations.EquationProblem, which can for example be an GeometricEquations.ODEProblem or GeometricEquations.PODEProblem. The optional parameter step determines the intervals for storing the solution, i.e., if step > 1 only every step'th solution is actually stored.