GeometricSolutions

Documentation for GeometricSolutions.

GeometricSolutions.GeometricSolution

GeometricSolutions.GeometricSolution — Type

GeometricSolution: Solution of an ordinary differential equation

Contains all fields necessary to store the solution of an ODE.

Fields

nt: number of time steps to store

t: time steps

s: NamedTuple of DataSeries for each solution component

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

Constructors

GeometricSolution(problem; nsave=DEFAULT_NSAVE)
GeometricSolution(t::TimeSeries, q::DataSeries, ntimesteps)

The usual way to initialise a Solution is by passing an equation, which for GeometricSolution has to be an ODEProblem or SODEProblem, a time step Δt and the number of time steps ntimesteps. The optional parameters nsave and nwrite determine the intervals for storing the solution and writing to file, i.e., if nsave > 1 only every nsave'th solution is actually stored, and every nwrite'th time step the solution is stored to disk.

The other constructors, either passing a TimeSeries and a DataSeries or a filename are used to read data from previous simulations.